# specs/003-fundaciones-datos/contracts/openapi-delta.yaml
openapi: 3.1.0
info:
  title: Descanso API — cambios de la feature 003
  version: 1.2.0
  description: |
    Cambios sobre specs/001-linea-base/contracts/openapi.yaml y
    specs/002-pipeline-ci-cd/contracts/openapi-delta.yaml. Todo lo demás no cambia.
    Todos los cambios son compatibles hacia atrás para clientes correctos: solo se añaden campos
    de respuesta y se rechazan entradas que antes se guardaban mal o daban 500.
servers:
  - url: /api
paths:
...
  /sleep:
    get:
      tags: [sleep]
      summary: Lista de noches — from/to validados (DT-03, FR-014)
      parameters:
        - $ref: '#/components/parameters/FromOptional'
        - $ref: '#/components/parameters/ToOptional'
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [sleep]
      summary: Crear noche — date debe ser el día local de bedtime (DT-04, FR-015)
      responses:
        '400':
          description: Además de los casos de 001, `date` distinta del día local de `bedtime`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'La fecha de la noche debe ser 2026-09-05 (el día en que te acostaste)' }
        '409':
          description: |
            Ya hay una noche abierta (sin cambios de mensaje). Ahora también lo garantiza la base
            con un índice único parcial (FR-019).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'Ya hay una noche abierta. Ciérrala antes de abrir otra.' }
  /sleep/{id}:
    put:
      tags: [sleep]
      summary: Editar noche — mismas reglas de date y noche abierta que POST
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  parameters:
    FromOptional:
      name: from
      in: query
      required: false
      schema: { type: string, format: date }
      description: Fecha real de calendario AAAA-MM-DD (se rechaza 2026-02-30). Debe ser ≤ to.
    ToOptional:
      name: to
      in: query
      required: false
      schema: { type: string, format: date }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Mensaje en español para el usuario }
  responses:
    BadRequest:
      description: Petición inválida (nunca 500 por parámetros mal formados)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            rango: { value: { error: 'from y to son obligatorios (AAAA-MM-DD)' } }
            fecha: { value: { error: 'from debe ser una fecha AAAA-MM-DD válida' } }
            orden: { value: { error: 'from no puede ser posterior a to' } }
    Conflict:
      description: Ya hay una noche abierta
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }